llm_security_gateway/backend/src/backend/api/middleware/logging.py
"""Request/response logging middleware."""

import time
import uuid
from typing import Callable

from fastapi import Request, Response
from starlette.middleware.base import BaseHTTPMiddleware

from backend.config.settings import get_settings


class RequestLoggingMiddleware(BaseHTTPMiddleware):
    """Middleware for logging requests and responses."""

    async def dispatch(self, request: Request, call_next: Callable) -> Response:
        """Log request and response details."""
        settings = get_settings()

        # Generate request ID
        request_id = str(uuid.uuid4())
        request.state.request_id = request_id

        # Record start time
        start_time = time.time()

        # Log request
        if settings.log_requests:
            print(
                f"[{request_id}] {request.method} {request.url.path} - Started"
            )

        # Process request
        response = await call_next(request)

        # Calculate duration
        duration = time.time() - start_time

        # Log response
        if settings.log_responses:
            print(
                f"[{request_id}] {request.method} {request.url.path} - "
                f"{response.status_code} ({duration:.3f}s)"
            )

        # Add request ID to response headers
        response.headers["X-Request-ID"] = request_id
        response.headers["X-Response-Time"] = f"{duration:.3f}s"

        return response
